refactor(admin): extract filter-panel attr catalog to keep it under 500 lines
advanced-filter-panel.tsx crossed the admin max-lines guard (513 > 500),
pushing the over-limit count to 22 vs the frozen baseline of 21 and
red-lighting every PR that triggers the frontend checks. Move the pure
static attribute catalog (PanelAttr, PANEL_ATTRS, the system/first attrs)
and the toFilterType mapping into a sibling advanced-filter-panel-attrs.ts
with no behaviour change; the panel imports them back. Count returns to 21.
