+++ apps/admin/src/components/catalog/advanced-filter-panel.tsx
@@ -16,6 +16,14 @@ import {
 import { jsonFetch } from '@/lib/http';
 import { cn } from '@/lib/utils';
 
+import {
+  FIRST_PANEL_ATTR,
+  PANEL_ATTRS,
+  type PanelAttr,
+  SYSTEM_PANEL_ATTRS,
+  toFilterType,
+} from './advanced-filter-panel-attrs';
+
 /**
  * VIEW-09 — push-down sticky-collapsible advanced filter panel.
  * Replaces Sheet-based `AdvancedFilterBuilder` (UI-02.9 / #299).
@@ -26,81 +34,7 @@ import { cn } from '@/lib/utils';
  * (BE base64 blob) that has to be maintained.
  */
 
-export type PanelAttr = {
-  code: string;
-  name: string;
-  type: keyof typeof FILTER_OPERATORS_BY_TYPE;
-  star?: boolean;
-};
-
-const FIRST_PANEL_ATTR: PanelAttr = {
-  code: 'brand',
-  name: 'Marka',
-  type: 'relation',
-  star: true,
-};
-
-/**
- * DASH-01 (#2249) — system list columns that are filterable in the search
- * index (IndexSettingsTemplate) but are not Attribute entities, so the
- * live `/api/attributes` fetch never returns them. They are unioned into
- * the effective catalog below; without this a condition seeded on
- * `completeness_pct` (dashboard deep-links) rendered with text operators
- * once the live set replaced the static fallback.
- */
-const SYSTEM_PANEL_ATTRS: ReadonlyArray<PanelAttr> = [
-  { code: 'completeness_pct', name: 'Completeness %', type: 'number', star: true },
-  { code: 'enabled', name: 'Aktywny', type: 'boolean', star: true },
-];
-
-/**
- * Loading/empty fallback catalog. Since #1354 the live filterable
- * attribute set (fetched below) is the source of truth for type-badge /
- * operator inference; this static list only fills the gap while that
- * fetch is in flight (or returns nothing) so the panel never renders a
- * dead picker on first open. An explicit `panelAttrs` prop still wins.
- */
-const PANEL_ATTRS: ReadonlyArray<PanelAttr> = [
-  FIRST_PANEL_ATTR,
-  { code: 'category', name: 'Kategoria', type: 'relation', star: true },
-  ...SYSTEM_PANEL_ATTRS,
-  { code: 'price', name: 'Cena bazowa', type: 'metric', star: true },
-  { code: 'stock', name: 'Stan magazynowy', type: 'number' },
-  { code: 'main_image', name: 'Główne zdjęcie', type: 'asset' },
-  { code: 'description.pl', name: 'Opis · PL', type: 'text' },
-  { code: 'description.en', name: 'Opis · EN', type: 'text' },
-  { code: 'meta_description', name: 'Meta description', type: 'text' },
-  { code: 'tags', name: 'Tagi', type: 'multiselect' },
-];
-
-/**
- * #1354 — map the backend AttributeType enum onto the panel's operator
- * buckets. Types without a dedicated bucket (color, email, identifier,
- * textarea) behave as plain string equality filters → `text`.
- */
-const FILTER_TYPE_BY_ATTR_TYPE: Record<string, keyof typeof FILTER_OPERATORS_BY_TYPE> = {
-  text: 'text',
-  wysiwyg: 'wysiwyg',
-  textarea: 'text',
-  number: 'number',
-  metric: 'metric',
-  price: 'price',
-  date: 'date',
-  datetime: 'datetime',
-  select: 'select',
-  multiselect: 'multiselect',
-  boolean: 'boolean',
-  relation: 'relation',
-  reference: 'reference',
-  asset: 'asset',
-  color: 'text',
-  email: 'text',
-  identifier: 'text',
-};
-
-function toFilterType(attrType: string | undefined): keyof typeof FILTER_OPERATORS_BY_TYPE {
-  return (attrType !== undefined ? FILTER_TYPE_BY_ATTR_TYPE[attrType] : undefined) ?? 'text';
-}
+export type { PanelAttr };
 
 /**
  * #2312 — Polish labels for the comparison operators. The select used to
